refactor: TEE attestation verification API
