+++ README.md
@@ -703,7 +703,7 @@ const { isVerified } = await verifyProof(proof, {
 
 ## TEE Attestation Verification
 
-The SDK supports verifying TEE (Trusted Execution Environment) attestations included in proofs. The current attestation flow verifies the Google Confidential Computing OIDC token returned by Popcorn's GCP attestor and checks that it is bound to the proof nonce and image digests.
+The SDK supports verifying TEE (Trusted Execution Environment) attestations included in proofs. The attestation flow verifies the Google Confidential Computing OIDC token returned by Popcorn's GCP attestor and checks that it is bound to the proof nonce, application identity, and image digests.
 
 ### Enabling TEE Attestation
 
@@ -715,9 +715,9 @@ const proofRequest = await ReclaimProofRequest.init(APP_ID, APP_SECRET, PROVIDER
 });
 ```
 
-### Verifying TEE Attestation
+### Verifying TEE Attestation via `verifyProof`
 
-Provide a `teeAttestation` config to require and verify TEE attestation. If TEE data is missing or invalid, verification will fail with a `TeeVerificationError`.
+Provide a `teeAttestation` config to require and verify TEE attestation. The application ID is derived automatically from `appSecret`. If TEE data is missing or invalid, verification will fail with a `TeeVerificationError`.
 
 ```javascript
 import { verifyProof, TeeVerificationError } from "@reclaimprotocol/js-sdk";
@@ -730,8 +730,8 @@ const { isVerified, isTeeVerified, data, error } = await verifyProof(proof, {
 });
 
 if (isVerified) {
-  console.log("Proof is fully verified with hardware attestation");
-  console.log("TEE verified:", isTeeVerified);
+  console.log("Proof verified with hardware attestation");
+  console.log("TEE verified:", isTeeVerified); // always true when teeAttestation is provided and passes
   console.log("Extracted parameters:", data[0].extractedParameters);
 } else if (error instanceof TeeVerificationError) {
   console.log("TEE verification failed:", error.message);
@@ -740,57 +740,53 @@ if (isVerified) {
 }
 ```
 
-The result always includes `isTeeVerified`. It is `true` only when `teeAttestation` was provided and verification passed; `false` otherwise. The overall `isVerified` will be `false` if TEE data is missing or TEE verification fails.
+The result always includes `isTeeVerified` as a boolean. It is `true` only when `teeAttestation` was provided and verification passed; `false` otherwise.
 
-The TEE verification validates:
+### What TEE verification checks
+
+- **Application binding**: Derives the application ID from `appSecret` and confirms the attestation was generated for your application
 - **Nonce binding**: Ensures the attestation nonce matches the proof context
-- **Application ID**: Confirms the attestation was generated for your application
 - **Session and timestamp binding**: Verifies the nonce metadata matches the proof session and is within the allowed skew
-- **OIDC token signature**: Validates the Google Confidential Computing attestation JWT against Google's JWKS (cached for 5 minutes)
-- **Platform claims**: Confirms the expected GCP confidential-computing claims such as issuer, secure boot, hardware model, and GCE instance metadata
+- **OIDC token signature**: Validates the Google Confidential Computing attestation JWT against Google's JWKS (responses are cached for 5 minutes)
+- **Platform claims**: Confirms the expected GCP confidential-computing claims (issuer, secure boot, hardware model, GCE instance metadata)
 - **Digest binding**: Confirms the workload and verifier image digests are present in the attestation token nonce list
 
-For hash-based attestation nonces, pass your **Application Secret** as `appSecret` so the SDK can recompute the expected nonce locally.
+### Standalone `verifyTeeAttestation`
 
-You can also verify TEE attestation separately using the standalone `verifyTeeAttestation` function:
+You can verify TEE attestation for a single proof without running full proof verification:
 
 ```javascript
 import { verifyTeeAttestation } from "@reclaimprotocol/js-sdk";
 
-const { isVerified, error } = await verifyTeeAttestation(proof, APP_ID, APP_SECRET);
+const { isVerified, error } = await verifyTeeAttestation(proof, APP_SECRET);
 if (isVerified) {
-  console.log("TEE attestation verified — proof was generated in a secure enclave");
+  console.log("TEE attestation verified");
 } else {
   console.log("TEE verification failed:", error);
 }
 ```
 
+`appSecret` is your Reclaim application secret. The application ID is derived from it automatically.
+
 ### Browser vs Server Verification
 
-TEE verification validates the Google Confidential Computing attestation JWT by fetching Google's OIDC metadata and JWKS. In browser-only environments this may fail due to cross-origin restrictions when fetching Google's discovery endpoints.
+TEE verification fetches Google's OIDC metadata and JWKS endpoints. In browser-only environments this may fail due to cross-origin restrictions.
 
-For web apps, prefer verifying TEE attestations on the server:
+For web apps, verify TEE attestations on the server:
 
 ```javascript
-// POST your proof(s) to your server and verify there
 const response = await fetch('/api/verify-tee', {
   method: 'POST',
   headers: { 'Content-Type': 'application/json' },
-  body: JSON.stringify({
-    proofs,
-    expectedApplicationId: APP_ID,
-    applicationSecret: APP_SECRET,
-  }),
+  body: JSON.stringify({ proofs }),
 });
 
 const { results } = await response.json();
 console.log(results);
 ```
 
-The example app included in this repository now uses a server route for TEE verification:
+The server route should read the app secret from environment variables only (never accept it from the client request body). See the example app for a reference implementation:
 - `example/src/app/api/verify-tee/route.ts`
-
-That route returns per-proof TEE verification results and failure reasons so the UI can show why verification failed.
 
 ## Error Handling
