+++ src/index.ts
@@ -11,7 +11,7 @@ export * from './utils/sessionUtils';
 export type * from './utils/sessionUtils';
 export * from './witness';
 export type * from './witness';
-export { verifyTeeAttestation } from './utils/verifyTee';
+export { verifyTeeAttestation, runTeeVerification } from './utils/verifyTee';
 export type { TeeVerificationResult } from './utils/verifyTee';
 export { TeeVerificationError } from './utils/errors';
 // Export device detection utilities for debugging (optional)
